+++ lib/pages/create_result_page.dart
@@ -12,6 +12,8 @@ import '../features/toggle_button.dart';
 import '../models/member.dart';
 import '../models/result.dart';
 import '../repositories/auth/auth_repository_impl.dart';
+import '../utils/app/result_format.dart';
+import '../utils/app/result_types.dart';
 import '../utils/async_value_error_dialog.dart';
 import '../utils/constants/app_colors.dart';
 import '../utils/constants/measure.dart';
@@ -19,8 +21,6 @@ import '../utils/constants/member.dart';
 import '../utils/dialog.dart';
 import '../utils/json_converters/union_timestamp.dart';
 import '../utils/loading.dart';
-import '../utils/result_format.dart';
-import '../utils/result_types.dart';
 import '../utils/scaffold_messenger_service.dart';
 import '../utils/text_styles.dart';
 import '../widgets/app_over_scroll_indicator.dart';
