+++ lib/pages/get_started/src/get_started.dart
@@ -1,0 +1,1 @@
+export '../src/widgets/app_title_text.dart';
