+++ lib/pages/log_in/src/log_in.dart
@@ -1,4 +1,3 @@
 export '../src/widgets//password_text_form.dart';
 export '../src/widgets/email_text_form.dart';
 export '../src/widgets/forget_password_text_button.dart';
-export '../src/widgets/gradation_background.dart';
