+++ lib/utils/app/result_types.dart
